+++ NEWS.md
@@ -1,3 +1,5 @@
+- added `only = FALSE` argument to `check_vector()` to check whether 
+only the actual values are permitted.
 - added `check_rbind()` to check two data frames can be smoothly `rbind`ed
 
 # checkr 0.2.0